Extract shared Python package to eliminate column_mapper duplication - Closes #33 - Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
