+++ api/Dockerfile
@@ -2,10 +2,11 @@ FROM python:3.13-slim
 
 WORKDIR /app
 
-COPY pyproject.toml .
+COPY shared/ /shared/
+COPY api/pyproject.toml .
 RUN pip install --no-cache-dir .
 
-COPY . .
+COPY api/ .
 
 EXPOSE 8000
 CMD ["uvicorn", "src.main:app", "--host", "0.0.0.0", "--port", "8000"]
