+++ api/pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "celery[redis]>=5.4",
     "httpx>=0.28",
     "redis>=5.0",
+    "openhedgefund-common @ file:///../shared",
 ]
 
 [project.optional-dependencies]
