+++ jobs/Dockerfile
@@ -2,7 +2,8 @@ FROM python:3.13-slim
 
 WORKDIR /app
 
-COPY pyproject.toml .
+COPY shared/ /shared/
+COPY jobs/pyproject.toml .
 RUN pip install --no-cache-dir .
 
-COPY . .
+COPY jobs/ .
