+++ shared/pyproject.toml
@@ -1,0 +1,10 @@
+[project]
+name = "openhedgefund-common"
+version = "0.1.0"
+description = "Shared utilities for OpenHedgeFund services"
+requires-python = ">=3.12"
+dependencies = []
+
+[tool.ruff]
+target-version = "py312"
+line-length = 100
